#5291 Create a special API to delete all objects with the same group (#5304)
This commit adds an new "private" API to Configurator. The purpose of new API is used to quickly delete ohara components from store. Also, the deletion does not include any safely check so it is dangerous to call this new API if you don't understand the details. Overall, this API should work for UI only and we will implement the new batch deletion APIs to replace this private and unsafe one.

Reviewers: Chih-Chiang Yeh <eechih@gmail.com>