#5581 Enable users to configure the compression type used by source c… (#5582)
This patch enables users to configure the compression type of source connector. The compression is a powerful tool. Users can leverage compression to save the I/O bandwidth to get higher performance. However, we don't enable it by default as the side effect of compression is the higher CPU usage. This tradeoff is up to users.

The new config is producer.compression.type

Reviewers: jackyoh <jack@is-land.com.tw>